SCP-5029 create common file for web e2e tests
