+++ marlowe-integration-tests/marlowe-integration-tests.cabal
@@ -64,6 +64,7 @@ executable marlowe-integration-tests
     Language.Marlowe.Runtime.Integration.MarloweQuery
     Language.Marlowe.Runtime.IntegrationSpec
     Language.Marlowe.Runtime.WebSpec
+    Language.Marlowe.Runtime.Web.Common
     Language.Marlowe.Runtime.Web.GetContracts
     Language.Marlowe.Runtime.CliSpec
   build-depends:
